Write bubble_sort method and style with rubocop
It was mostly a matter of first choosing the methods to use, then applying them meticulously so it all makes sense. Made some oopsies left and right but got there in the end.

I know the destructive methods aren't preferred, but it made this whole ordeal a lot more concise.